Function dft::real::forward

pub fn forward(data: &mut [f64])

Perform the forward transform.

The number of points should be a power of two. The data are replaced by the positive frequency half of their complex Fourier transform. The real-valued first and last components of the complex transform are returned as elements data[0] and data[1], respectively.

References

William H. Press, Saul A. Teukolsky, William T. Vetterling, Brian P. Flannery, “Numerical Recipes 3rd Edition: The Art of Scientific Computing,” Cambridge University Press, 2007.